requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-Migrate==1.4.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
flask-ldap3-login==0.9.7
grequests==0.2.0
jsmin==2.1.1
requests==2.6.2
pexpect==3.3
pytest==2.7.0
pytest-cov==1.8.1
supervisor>=3.1.3
uWSGI==2.0.13
